==> 2b_random_forest.Rmd <==

---
title: "Random_Forest_Project2"
author: "Nathan Koh"
date: "2024-12-03"
output: html_document
---

```{r setup, include=FALSE}
# Setting up knitr options to display code and output in the document
knitr::opts_chunk$set(echo = TRUE)
```

## Library

```{r}
# Importing the randomForest library for building and evaluating Random Forest models
library(randomForest)
library(caret)
```

## Step 1: Import Data

```{r}
# Reading in the cleaned hotel booking dataset
hotel <- read.csv("cleaned.csv")

# Displaying the first few rows of the dataset for a quick preview
head(hotel)

# Displaying the structure of the dataset to check variable types and dimensions
str(hotel)
```

## Step 2: Clean Data

```{r}
# Removing the unnecessary 'X' column (index column)
hotel$X <- NULL

# Converting all categorical variables into dummy variables using model.matrix
hotel_dummy <- as.data.frame(model.matrix(~.-1, data = hotel))

# Defining a min-max scaling function to scale numeric data between 0 and 1
min_max <- function(x) {
  (x - min(x)) / (max(x) - min(x))
}

# Applying min-max scaling to all columns in the dataset to standardize the data
hotel_scaled <- as.data.frame(lapply(hotel_dummy, min_max))

# Summarizing the scaled dataset to confirm the transformations
summary(hotel_scaled)
```

## Step 3: Split Data

```{r}
# Setting a seed for reproducibility when splitting the data
set.seed(12345)

# Defining the proportion of data to be used for testing
test_size <- 0.3

# Randomly sampling row indices for the test set based on the test size
test_rows <- sample(1:nrow(hotel_scaled), test_size * nrow(hotel_scaled))

# Splitting the data into test and train sets
hotel_test <- hotel_scaled[test_rows, ]  # Test dataset
hotel_train <- hotel_scaled[-test_rows, ]  # Training dataset

# Ensuring the 'booking.status' variable is treated as a factor for classification
hotel_train$booking.status <- as.factor(hotel_train$booking.status)
hotel_test$booking.status <- as.factor(hotel_test$booking.status)
```

## Step 4: Random Forest Model

```{r, cache = TRUE}
# Building a Random Forest model to predict 'booking.status' using all other variables
# ntree specifies the number of trees in the forest, and mtry controls the number of variables randomly sampled at each split
m1 <- randomForest(booking.status ~ ., data = hotel_train, ntree = 500, mtry = sqrt(18))

# Plotting the error rates for the Random Forest model as the number of trees increases
plot(m1)

# Generating predictions for the test set
pred1 <- predict(m1, hotel_test, type = "class")

# Displaying the variable importance plot to identify the most influential predictors
randomForest::varImpPlot(m1)

# Evaluating the model's performance using a confusion matrix
confusionMatrix(table(pred1, hotel_test$booking.status))
```

## Step 5: Export Predictions

```{r}
# Exporting the predictions to a CSV file for further analysis or reporting
write.csv(pred1, "prediction_random_forest.csv")
```
